✅ D9VQVH code: done
Task: 202604260755-D9VQVH
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: Removed obsolete internal testing and shim surfaces, moved remaining consumers to owned modules or testkit, and verified with typecheck, knip, focused tests, formatting, bootstrap, and doctor.
