--- packages/agentplane/src/commands/pr/integrate/internal/cleanup.ts
@@ -1,1 +1,0 @@
-export { cleanupMergedLocalBranch as cleanupIntegratedBranch } from "../../../shared/merged-branch-cleanup.js";
