+++ packages/agentplane/src/commands/release/write-release-ready-manifest-script.test.ts
@@ -5,8 +5,8 @@ import { promisify } from "node:util";
 
 import { afterEach, describe, expect, it } from "vitest";
 
+import { writeExecutableFile } from "@agentplane/testkit";
 import { initReleaseWorkspace } from "@agentplane/testkit/release";
-import { writeExecutableFile } from "../../test-helpers/fs.js";
 
 const execFileAsync = promisify(execFile);
 const SCRIPT_PATH = path.resolve(process.cwd(), "scripts/manifest.mjs");
