--- packages/agentplane/src/runner/run-id.ts
@@ -1,3 +1,0 @@
-export function createRunnerRunId(date: Date = new Date()): string {
-  return date.toISOString().replaceAll(":", "-").replaceAll(".", "-");
-}
