+++ packages/agentplane/src/runtime/incidents/index.ts
@@ -2,21 +2,13 @@ export type {
   IncidentAdviceMatch,
   IncidentAdviceQuery,
   IncidentCollectionPlan,
-  IncidentFindingCandidate,
-  IncidentSkippedFinding,
-  IncidentPromotionDraft,
-  IncidentPromotionIssue,
-  IncidentPromotionTaskContext,
   IncidentRegistry,
-  IncidentRegistryEntry,
-  IncidentRegistryEntryState,
 } from "./types.js";
+export { INCIDENTS_POLICY_ASSET_PATH, INCIDENTS_POLICY_PATH } from "./types.js";
 export {
   appendIncidentRegistryEntries,
   buildIncidentAdviceQueryFromTask,
   createIncidentRegistrySkeleton,
-  extractIncidentCandidatesFromFindings,
-  formatIncidentRegistryEntry,
   parseIncidentRegistry,
   planIncidentCollection,
   renderIncidentAdvice,
