--- packages/agentplane/src/runtime/incidents/paths.ts
@@ -1,2 +1,0 @@
-export const INCIDENTS_POLICY_PATH = ".agentplane/policy/incidents.md";
-export const INCIDENTS_POLICY_ASSET_PATH = "packages/agentplane/assets/policy/incidents.md";
