+++ packages/agentplane/src/runtime/incidents/types.ts
@@ -1,3 +1,6 @@
+export const INCIDENTS_POLICY_PATH = ".agentplane/policy/incidents.md";
+export const INCIDENTS_POLICY_ASSET_PATH = "packages/agentplane/assets/policy/incidents.md";
+
 export type IncidentRegistryEntryState = "open" | "stabilized" | "promoted";
 
 export type IncidentRegistryEntry = {
