--- packages/agentplane/src/testing/index.ts
@@ -1,20 +1,0 @@
-export { runCli } from "../cli/run-cli.js";
-export { infoMessage } from "../cli/output.js";
-export {
-  filterAgentsByWorkflow,
-  loadAgentTemplates,
-  loadAgentsTemplate,
-} from "../agents/agents-template.js";
-export type * as taskBackend from "../backends/task-backend.js";
-export { loadCommandContext, loadTaskFromContext } from "../commands/shared/task-backend.js";
-export { evolveRunnerRunState, writeRunnerRunState } from "../runner/artifacts.js";
-export {
-  formatRunnerCapabilitySummaryLines,
-  formatRunnerPolicyFieldSummaryLines,
-} from "../runner/policy-display.js";
-export type { RunnerAdapterCapabilities, RunnerPolicyDecision } from "../runner/types.js";
-export { prepareTaskRunnerExecution } from "../runner/usecases/task-run.js";
-export * as processSupervision from "../runner/process-supervision/signals.js";
-export { resolveUpdateCheckCachePath } from "../cli/update-check.js";
-export * as prompts from "../cli/prompts.js";
-export { VERIFY_STEPS_PLACEHOLDER } from "../commands/task/shared/docs.js";
