+++ packages/testkit/src/agentplane-internal.ts
@@ -1,0 +1,29 @@
+export { runCli } from "../../agentplane/src/cli/run-cli.js";
+export { infoMessage } from "../../agentplane/src/cli/output.js";
+export {
+  filterAgentsByWorkflow,
+  loadAgentTemplates,
+  loadAgentsTemplate,
+} from "../../agentplane/src/agents/agents-template.js";
+export type * as taskBackend from "../../agentplane/src/backends/task-backend.js";
+export {
+  loadCommandContext,
+  loadTaskFromContext,
+} from "../../agentplane/src/commands/shared/task-backend.js";
+export {
+  evolveRunnerRunState,
+  writeRunnerRunState,
+} from "../../agentplane/src/runner/artifacts.js";
+export {
+  formatRunnerCapabilitySummaryLines,
+  formatRunnerPolicyFieldSummaryLines,
+} from "../../agentplane/src/runner/policy-display.js";
+export type {
+  RunnerAdapterCapabilities,
+  RunnerPolicyDecision,
+} from "../../agentplane/src/runner/types.js";
+export { prepareTaskRunnerExecution } from "../../agentplane/src/runner/usecases/task-run.js";
+export * as processSupervision from "../../agentplane/src/runner/process-supervision/signals.js";
+export { resolveUpdateCheckCachePath } from "../../agentplane/src/cli/update-check.js";
+export * as prompts from "../../agentplane/src/cli/prompts.js";
+export { VERIFY_STEPS_PLACEHOLDER } from "../../agentplane/src/commands/task/shared/docs.js";
