+++ packages/testkit/src/fixtures.ts
@@ -1,4 +1,5 @@
-import { rm } from "node:fs/promises";
+import { mkdir, rm, writeFile } from "node:fs/promises";
+import path from "node:path";
 
 import { defaultConfig } from "@agentplaneorg/core/config";
 
@@ -58,4 +59,16 @@ export function mockConfig(
   const config = defaultConfig();
   configure?.(config);
   return config;
+}
+
+export async function writeExecutableFile(
+  root: string,
+  relativePath: string,
+  content: string | readonly string[],
+): Promise<string> {
+  const target = path.join(root, relativePath);
+  const body = Array.isArray(content) ? content.join("\n") : String(content);
+  await mkdir(path.dirname(target), { recursive: true });
+  await writeFile(target, `${body}\n`, { encoding: "utf8", mode: 0o755 });
+  return target;
 }
